requirements.txt
# Python dependencies for Batch Scanning Jig
# Install with: pip3 install -r requirements.txt

pyserial>=3.5
RPLCD>=1.3.0
